Story refactors, Moved redundant components to temp old folder
