+++ old/banner/src/atoms/info-banner.tsx
