+++ old/summary/src/atoms/redacted-description.tsx
