+++ packages/storybook/src/organisms/styled/card/card.stories.tsx
@@ -3,11 +3,9 @@ import { storiesOf } from '@storybook/react'
 import JSXAddon from 'storybook-addon-jsx'
 
 import readme from '@ltht-react/card/README.md'
-import { ListItem } from '@ltht-react/list'
-import { InfoSummary, WarningSummary, ErrorSummary } from '@ltht-react/summary'
 import Card from '@ltht-react/card'
 import AllergySummary from '@ltht-react/allergy-summary'
-import { CounterIcon } from '@ltht-react/icon'
+import { CounterIcon, ExclamationIcon, InfoCircleIcon } from '@ltht-react/icon'
 import allergies from '../../clinical/allergies/allergy.fixtures'
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -32,125 +30,37 @@ stories.addWithJSX('Basic', () => (
       <Card.Header>
         <Card.Title>Header</Card.Title>
       </Card.Header>
-      <Card.Banner status="warning" onClick={clickHandler}>
-        Hello
-      </Card.Banner>
       <Card.Body>
         <Card.Subtitle>Subtitle</Card.Subtitle>
         <Card.Text>Body Text</Card.Text>
       </Card.Body>
     </Card>
-    <Card>
-      <Card.Header>
-        <Card.Title>Header</Card.Title>
-      </Card.Header>
-      <Card.List>
-        <Card.ListItem>
-          <Card.Text>Item One</Card.Text>
-        </Card.ListItem>
-        <Card.ListItem>
-          <Card.Text>Item Two</Card.Text>
-        </Card.ListItem>
-        <Card.ListItem>
-          <Card.Text>Item Three</Card.Text>
-        </Card.ListItem>
-      </Card.List>
-    </Card>
-    <Card>
-      <Card.Header>
-        <Card.Title>Header</Card.Title>
-      </Card.Header>
-      <Card.Alert>Default Alert</Card.Alert>
-      <Card.List>
-        <Card.ListItem>Item One</Card.ListItem>
-        <Card.ListItem>Item Two</Card.ListItem>
-        <Card.ListItem>Item Three</Card.ListItem>
-      </Card.List>
-    </Card>
-    <Card>
-      <Card.Header>
-        <Card.Title>Header</Card.Title>
-      </Card.Header>
-      <Card.Alert onClick={clickHandler} status="info">
-        Info Alert
-      </Card.Alert>
-      <Card.List>
-        <Card.ListItem>Item One</Card.ListItem>
-        <Card.ListItem>Item Two</Card.ListItem>
-        <Card.ListItem>Item Three</Card.ListItem>
-      </Card.List>
-    </Card>
-    <Card>
-      <Card.Header>
-        <Card.Title>Header</Card.Title>
-      </Card.Header>
-      <Card.Alert status="danger">Danger Alert</Card.Alert>
-      <Card.Body>
-        <Card.Text>Body Text</Card.Text>
-      </Card.Body>
-      <Card.List>
-        <Card.ListItem>Item One</Card.ListItem>
-        <Card.ListItem>Item Two</Card.ListItem>
-        <Card.ListItem>Item Three</Card.ListItem>
-      </Card.List>
-    </Card>
-    <Card>
-      <Card.Header>
-        <Card.Title>Header</Card.Title>
-      </Card.Header>
-      <Card.Alert status="warning">Warning Alert</Card.Alert>
-      <Card.List>
-        <Card.ListItem>Item One</Card.ListItem>
-        <Card.ListItem>Item Two</Card.ListItem>
-        <Card.ListItem>Item Three</Card.ListItem>
-      </Card.List>
-    </Card>
-    <Card>
-      <Card.List>
-        <Card.ListItem>Item One</Card.ListItem>
-        <Card.ListItem>Item Two</Card.ListItem>
-        <Card.ListItem>Item Three</Card.ListItem>
-      </Card.List>
-    </Card>
-    <Card>
-      <Card.Body>
-        <Card.Text>Text One</Card.Text>
-        <Card.Text>Text Two</Card.Text>
-        <Card.Text>Text Three</Card.Text>
-      </Card.Body>
-    </Card>
-    <Card>
-      <Card.Header>Header</Card.Header>
-      <Card.Alert status="warning">This is an alert</Card.Alert>
-      <Card.Body>Body</Card.Body>
-      <Card.Footer>Footer</Card.Footer>
-    </Card>
   </>
 ))
 
-stories.addWithJSX('Footer', () => (
+stories.addWithJSX('With Footer', () => (
   <Card>
     <Card.Header>
       <Card.Title>Header</Card.Title>
     </Card.Header>
-    <Card.Body>Body</Card.Body>
+    <Card.Body>
+      <Card.Subtitle>Subtitle</Card.Subtitle>
+      <Card.Text>Body Text</Card.Text>
+    </Card.Body>
     <Card.Footer>Footer</Card.Footer>
   </Card>
 ))
 
-stories.addWithJSX('List', () => (
+stories.addWithJSX('With List', () => (
   <div>
     <Card>
       <Card.Header>
         <Card.Title>Header</Card.Title>
       </Card.Header>
-      <Card.Body>
-        <Card.List>
-          <Card.ListItem>Item 1</Card.ListItem>
-          <Card.ListItem>Item 2</Card.ListItem>
-        </Card.List>
-      </Card.Body>
-      <Card.Footer>Footer</Card.Footer>
+      <Card.List>
+        <Card.ListItem>Item 1</Card.ListItem>
+        <Card.ListItem>Item 2</Card.ListItem>
+      </Card.List>
     </Card>
   </div>
 ))
@@ -161,13 +71,14 @@ stories.addWithJSX('List (Clickable)', () => (
       <Card.Header>
         <Card.Title>Header</Card.Title>
       </Card.Header>
-      <Card.Body>
-        <Card.List>
-          <Card.ListItem>Item 1</Card.ListItem>
-          <Card.ListItem>Item 2</Card.ListItem>
-        </Card.List>
-      </Card.Body>
-      <Card.Footer>Footer</Card.Footer>
+      <Card.List>
+        <Card.ListItem onClick={clickHandler}>
+          <div>Item 1</div>
+        </Card.ListItem>
+        <Card.ListItem onClick={clickHandler}>
+          <div>Item 2</div>
+        </Card.ListItem>
+      </Card.List>
     </Card>
   </div>
 ))
@@ -178,7 +89,9 @@ stories.addWithJSX('Info Summary', () => (
       <Card.Header>
         <Card.Title>Allergies</Card.Title>
       </Card.Header>
-      <InfoSummary clickHandler={clickHandler} />
+      <Card.Banner status="info" onClick={clickHandler}>
+        <InfoCircleIcon size="medium" status="info" /> View Data Sources
+      </Card.Banner>
       <Card.List>
         {allergies.map(allergy => (
           <Card.ListItem>
@@ -186,11 +99,15 @@ stories.addWithJSX('Info Summary', () => (
           </Card.ListItem>
         ))}
       </Card.List>
-      <InfoSummary clickHandler={clickHandler} />
+      <Card.Banner status="info" onClick={clickHandler}>
+        <InfoCircleIcon size="medium" status="info" /> View Data Sources
+      </Card.Banner>
       <Card.List>
         <Card.ListItem>Missing 10</Card.ListItem>
       </Card.List>
-      <WarningSummary clickHandler={clickHandler} />
+      <Card.Banner status="warning" onClick={clickHandler}>
+        <ExclamationIcon size="medium" status="default" /> Incomplete Data
+      </Card.Banner>
       <Card.List>
         <Card.ListItem>Missing 10</Card.ListItem>
         <Card.ListItem>Missing 10</Card.ListItem>
@@ -205,7 +122,9 @@ stories.addWithJSX('Warning Summary', () => (
       <Card.Header>
         <Card.Title>Allergies</Card.Title>
       </Card.Header>
-      <WarningSummary clickHandler={clickHandler} />
+      <Card.Banner status="warning" onClick={clickHandler}>
+        <ExclamationIcon size="medium" status="default" /> Incomplete Data
+      </Card.Banner>
       <Card.List>
         {allergies.map(allergy => (
           <Card.ListItem>
@@ -223,12 +142,12 @@ stories.addWithJSX('Error Summary', () => (
       <Card.Header>
         <Card.Title>Allergies</Card.Title>
       </Card.Header>
-      <Card.Body>
-        <Card.List>
-          <ErrorSummary clickHandler={clickHandler} />
-          <ListItem>No Data</ListItem>
-        </Card.List>
-      </Card.Body>
+      <Card.Banner status="danger" onClick={clickHandler}>
+        <ExclamationIcon size="medium" status="default" /> Regional Providers Unavailable
+      </Card.Banner>
+      <Card.List>
+        <Card.ListItem>No Data</Card.ListItem>
+      </Card.List>
     </Card>
   </div>
 ))
@@ -273,13 +192,6 @@ stories.addWithJSX('Missing Data Summary', () => (
         <Card.Text>Bradford</Card.Text>
       </Card.ListItem>
     </Card.List>
-
-    {/* <MissingDataSummary
-            missingData={['Humber']}
-            partialData={['York (2 missing)']}
-            completeData={['Hull And East Yorkshire', 'Rotherham']}
-            noData={['Bradford']}
-          /> */}
   </Card>
 ))
